Handle feedback, add ESlint
