+++ example/components/html/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  update: require('./update'),
+};
