+++ example/components/text/footer.js
@@ -1,7 +1,5 @@
 const React = require('react');
 
-const style = require('../style');
-
 module.exports = React.createClass({
     render() {
       return (
