+++ example/components/text/header.js
@@ -1,7 +1,5 @@
 const React = require('react');
 
-const style = require('../style');
-
 module.exports = React.createClass({
   propTypes: {
     name: React.PropTypes.string.isRequired
