--- example/textComponents/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-  update:  require('./update'),
-};
